+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/savings/domain/GSIMRepositoy.java
@@ -19,7 +19,6 @@
 
 package org.apache.fineract.portfolio.savings.domain;
 
-import java.math.BigDecimal;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
@@ -28,7 +27,7 @@ public interface GSIMRepositoy
 
     GroupSavingsIndividualMonitoring findOneByIsAcceptingChild(boolean acceptingChild);
 
-    GroupSavingsIndividualMonitoring findOneByIsAcceptingChildAndApplicationId(boolean acceptingChild, BigDecimal applicationId);
+    GroupSavingsIndividualMonitoring findOneByIsAcceptingChildAndApplicationId(boolean acceptingChild, Long applicationId);
 
     GroupSavingsIndividualMonitoring findOneByAccountNumber(String accountNumber);
 }
